upload

+++ Practise/lab week 2 number 3.c
@@ -1,14 +1,14 @@
 main()
 {
-    int a,b,c,sum,average;
+    float a,b,c,sum,average;
     printf("Enter subject 1 mark:");
-    scanf("%d",&a);
+    scanf("%f",&a);
     printf("Enter subject 2 mark:");
-    scanf("%d",&b);
+    scanf("%f",&b);
     printf("Enter subject 3 mark:");
-    scanf("%d",&c);
+    scanf("%f",&c);
     sum=a+b+c;
     average=(a+b+c)/3;
-    printf("The total is %d+%d+%d=%d",a,b,c,sum);
-    printf("\nAverage is %d",average);
+    printf("The total is %.0f+%.0f+%.0f=%.0f",a,b,c,sum);
+    printf("\nAverage is %.2f",average);
 }
